requirements.txt
fastapi
uvicorn[standard]
python-multipart
pyyaml
numpy
opencv-python
ultralytics